fix: code fixes and refactor.
- Decouple AuthToken: Removed the user reverse relationship to enforce direct model filtering and prevent architectural leaks.
- Restructure: Moved AbstractAuthToken to base/models for a more logical project hierarchy.
- Admin Update: Switched admin form fields to __all__ to automatically capture inherited fields from the User model (move unrequired fields to exclude to prevent test issues).
