+++ drf_authentify/forms.py
@@ -11,7 +11,8 @@ AuthToken = get_token_model()
 class AuthTokenAdminForm(forms.ModelForm):
     class Meta:
         model = AuthToken
-        fields = ["user", "auth_type", "context", "expires_at", "refresh_until"]
+        fields = "__all__"
+        exclude = ("access_token_hash", "refresh_token_hash", "last_refreshed_at")
 
     def clean(self) -> dict:
         cleaned_data = super().clean()
